04-analytics-engineering/pyproject.toml
[project]
name = "04-analytics-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "dbt-duckdb>=1.10.0",
]
